% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPTOMultiSimulator-package.R
\docType{package}
\name{BPTOMultiSimulator-package}
\alias{BPTOMultiSimulator}
\alias{BPTOMultiSimulator-package}
\title{BPTO Simulator}
\description{
This is a boilerplate for BPTO Simulator for multiple HB
  conjoint models. It works with Alternative Specific design, where the
  Alternatives are brands, and their only attribute is Price. The package
  supports the models with some alternatives having constant price
  (for details please visit Vignette). To build the simulator you need to
  provide data in proper format, and the package will figure out itself the
  how to build the interface. As of now, it gives the insight in the
  Share of Preference, Revenue, Channel Sales and Profit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mczyzj/BPTOMultiSimulator}
  \item Report bugs at \url{https://github.com/mczyzj/BPTOMultiSimulator/issues}
}

}
\author{
\strong{Maintainer}: Michał Czyż \email{m.czyz.j@gmail.com}

}
\keyword{internal}
